bash: Customize prompt
This prompt is simpler than the one generated by promptline but
implements all the critical features in way fewer lines.

When not running under a GUI terminal emulator, purposefully
disable all the fancy features to reduce the chances that things
will break at the worst possible time, such as when performing
rescue operations with console-only access.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>